ログイン時の履歴書上書きを「保存ボタン押下時のみ」に修正

+++ app/controllers/sessions_controller.rb
@@ -4,7 +4,7 @@ class SessionsController < ApplicationController
   def create
     user = find_or_create_from_auth_hash(auth_hash)
     log_in(user) if user
-    after_save_for_guest(user) if session['resume_uuid']
+    after_save_for_guest(user)
     redirect_to root_path
   end
 
@@ -28,10 +28,11 @@ class SessionsController < ApplicationController
   end
 
   def after_save_for_guest(user)
-    user.skincare_resume.destroy! if user.skincare_resume.present?
+    return unless session['resume_uuid']
+    return unless request.env.dig('omniauth.params', 'button') == 'save'
 
-    resume = SkincareResume.find_by(uuid: session['resume_uuid'])
-    resume.update!(user:)
+    user.skincare_resume&.destroy!
+    SkincareResume.find_by!(uuid: session['resume_uuid']).update!(user:)
     session.delete('resume_uuid')
   end
 end
